requirements.txt
pillow
pyinstaller
pandas